requirements.txt
Django==2.2
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-discover-runner==1.0
django-imagekit==4.0.2
django-role-permissions==2.2.1
pilkit==2.0
Pillow==6.0.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
